// lib/enums/eApgWglMaterialTypes.ts
export enum eApgWglMaterialTypes {
  LineMaterials = 'LineMats',
  WireframeMaterials = 'wireframe_mats',
  LambertMaterials = 'lambert_mats',
  PhongMaterials = 'phong_mats',
  PhongFlatMaterials = 'phong_flat_mats',
  PhongTranspMaterials = 'phong_transp_mats',
  PhongTexturedMaterials = 'phong_txtrd_mats',
  PhongBumpedMaterials = 'phong_bumped_mats',
  StdMaterials = 'std_mats',
  StdTexturedMaterials = 'std_txtrd_mats',
  StdNormaledMaterials = 'std_nrmled_mats'
}